+++ web/src/modules/travel-authorizations/components/AccommodationTypeSelect.vue
