+++ web/src/modules/travel-authorizations/pages/travel-authorization-read/SummaryHeaderPanel.vue
