+++ web/src/modules/travel-authorizations/pages/travel-authorization-read/details-tab/DetailsCard.vue
