+++ web/src/modules/travel-authorizations/pages/travel-form-edit/details-tab/ApprovalsFormCard.vue
