+++ web/src/modules/travel-authorizations/pages/travel-form-edit/details-tab/details-form-card/OneWayStopsSection.vue
@@ -91,10 +91,10 @@ import DatePicker from "@/components/Utils/DatePicker"
 import TimePicker from "@/components/Utils/TimePicker"
 import AccommodationTypeSelect, {
   ACCOMMODATION_TYPES,
-} from "@/modules/travel-form/components/AccommodationTypeSelect"
+} from "@/modules/travel-authorizations/components/AccommodationTypeSelect"
 import TravelMethodSelect, {
   TRAVEL_METHODS,
-} from "@/modules/travel-form/components/TravelMethodSelect"
+} from "@/modules/travel-authorizations/components/TravelMethodSelect"
 
 export default {
   name: "OneWayStopsSection",
