+++ web/src/modules/travel-authorizations/pages/travel-form-edit/estimate-tab/EstimateCreateDialog.vue
@@ -93,7 +93,7 @@ import { required } from "@/utils/validators"
 
 import CurrencyTextField from "@/components/Utils/CurrencyTextField"
 import DatePicker from "@/components/Utils/DatePicker"
-import ExpenseTypeSelect from "@/modules/travel-form/components/ExpenseTypeSelect"
+import ExpenseTypeSelect from "@/modules/travel-authorizations/components/ExpenseTypeSelect"
 
 import expensesApi from "@/apis/expenses-api"
