+++ web/src/store/index.js
@@ -8,7 +8,7 @@ import preapproved from "./preapproved";
 import traveldesk from "./traveldesk";
 import reports from "./reports";
 
-import travelForm from "@/modules/travel-form/store";
+import travelForm from "@/modules/travel-authorizations/store";
 
 Vue.use(Vuex);
